my_telegram_bot/requirements.txt
Flask==2.0.1
python-telegram-bot==13.7
